+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Version 1.2.1
 
 ## Quality of Life
-- Runs now get displayed witn their parent directory for better distinguishability
+- Runs now get displayed with their parent directory for better distinguishability
+- Increase plot font sizes.
 
 # Version 1.2